+++ test/bigtest/network/scenarios/default.js
@@ -3,5 +3,5 @@
 // default scenario is used during `yarn start --mirage`
 export default function defaultScenario(server) {
   server.createList('jobExecution', 20);
-  server.get('/metadata-provider/jobExecutions', (schema) => schema.jobExecutions.all());
+  server.get('/metadata-provider/jobExecutions', schema => schema.jobExecutions.all());
 }